Stand-alone Python module for PostgresColumn
