+++ plugins/postgres/dbt/adapters/postgres/__init__.py
@@ -1,7 +1,7 @@
 # these are mostly just exports, #noqa them so flake8 will be happy
 from dbt.adapters.postgres.connections import PostgresConnectionManager  # noqa
 from dbt.adapters.postgres.connections import PostgresCredentials
-from dbt.adapters.postgres.relation import PostgresColumn  # noqa
+from dbt.adapters.postgres.column import PostgresColumn  # noqa
 from dbt.adapters.postgres.relation import PostgresRelation  # noqa: F401
 from dbt.adapters.postgres.impl import PostgresAdapter
